+++ test/allocation.jl
@@ -65,6 +65,7 @@ ManifoldsBase.representation_size(::AllocManifold3) = (2, 3)
     @test number_eltype(([2], [3.0])) === Float64
     @test number_eltype(([2], [3])) === Int
     @test number_eltype(Any[[2.0], [3.0]]) === Float64
+    @test number_eltype(typeof([[1.0, 2.0]])) === Float64
 
     alloc2 = ManifoldsBase.allocate_result(AllocManifold2(), rand)
     @test alloc2 isa Matrix{Float64}