+++ src/ManifoldsBase.jl
@@ -1058,6 +1058,7 @@ export AbstractInverseRetractionMethod,
 # (d) Vector Transport Types
 export AbstractVectorTransportMethod,
     DifferentiatedRetractionVectorTransport,
+    EmbeddedVectorTransport,
     ParallelTransport,
     PoleLadderTransport,
     ProductVectorTransport,