+++ NEWS.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.15.6] 15/12/2023
+## [0.15.6] unreleased
+
+### Added
+
+* `EmbeddedVectorTransport` to use a vector transport in the embedding and a final projection.
 
 ### Fixed